update logger outptus
